refactor: define deploy constants
