--- deploy/modes.py
@@ -1,5 +1,0 @@
-"""
-Allowed deployment modes
-"""
-DEVELOPMENT_MODE = 'development'
-PRODUCTION_MODE = 'production'
